requirements.txt
requests
pillow
imageio
